added cjsx loader and ability to load multiple loaders at once
