+++ package.json
@@ -32,6 +32,7 @@
   },
   "devDependencies": {
     "async": "~0.9.0",
+    "coffee-react": "^4.0.0",
     "coffee-script": "~1.7.1",
     "grunt": "~0.4.5",
     "grunt-cli": "^0.1.13",
@@ -42,6 +43,7 @@
     "load-grunt-tasks": "~0.4.0",
     "mocha": "~1.20.1",
     "phantomjs": "1.8.2-0",
+    "react": "^0.13.3",
     "requirejs": "~2.1.14",
     "travis-cov": "*",
     "uglify-save-license": "~0.4.1"
