+++ src/node-loaders/coffee-script.js
@@ -1,54 +1,12 @@
-var fs = require("fs"),
-    path = require("path");
+var simpleNodeLoader = require('./simple-node-loader');
 
 module.exports = function(blanket) {
     var coffeeScript = require("coffee-script"),
         oldLoaderCS = require.extensions['.coffee'];
 
-    require.extensions['.coffee'] = function(localModule, filename) {
-        var antipattern = blanket.options("antifilter"),
-            pattern = blanket.options("filter"),
-            reporter_options = blanket.options("reporter_options"),
-            originalFilename = filename;
-        filename = blanket.normalizeBackslashes(filename);
-
-        if (typeof antipattern !== "undefined" && blanket.matchPattern(filename.replace(/\.coffee$/, ""), antipattern)) {
-
-            oldLoaderCS(localModule, filename);
-            if (blanket.options("debug")) {
-                console.log("BLANKET-File will never be instrumented:" + filename);
-            }
-        } else if (blanket.matchPattern(filename, pattern)) {
-            if (blanket.options("debug")) {
-                console.log("BLANKET-Attempting instrument of:" + filename);
-            }
-
-            var content = fs.readFileSync(filename, 'utf8');
-            content = coffeeScript.compile(content);
-
-            if (reporter_options && reporter_options.shortnames){
-                inputFilename = filename.replace(path.dirname(filename),"");
-            } else if (reporter_options && reporter_options.relativepath) {
-                inputFilename = filename.replace(process.cwd(),"");
-            }
-            if (reporter_options && reporter_options.basepath){
-                inputFilename = filename.replace(reporter_options.basepath + '/',"");
-            }
-
-            blanket.instrument({
-                inputFile: content,
-                inputFileName: inputFilename
-            }, function(instrumented) {
-                var baseDirPath = blanket.normalizeBackslashes(path.dirname(filename)) + '/.';
-                try {
-                    instrumented = instrumented.replace(/require\s*\(\s*("|')\./g, 'require($1' + baseDirPath);
-                    localModule._compile(instrumented, originalFilename);
-                } catch (err) {
-                    console.log("Error parsing instrumented code: " + err);
-                }
-            });
-        } else {
-            oldLoaderCS(localModule, filename);
-        }
+    var compile = function(content) {
+        return coffeeScript.compile(content);
     };
+
+    require.extensions['.coffee'] = simpleNodeLoader(blanket, oldLoaderCS, compile, /\.coffee$/);
 };
