+++ test/test-node/fixture/src/sampleCjsx.cjsx
@@ -1,0 +1,4 @@
+React = require 'react'
+
+module.exports = ->
+  <div>10</div>
