+++ test/test-node/testrunner_cjsx.js
@@ -1,0 +1,4 @@
+var blanket = require("../../src/index")();
+blanket.options("filter","src/sample");
+
+require("./tests/cjsx/test.coffee");
